+++ apps/client/src/app/app.module.ts
@@ -19,8 +19,8 @@ import { MatTooltipModule } from '@angular/material/tooltip';
 import { BrowserModule } from '@angular/platform-browser';
 import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
 import { ServiceWorkerModule } from '@angular/service-worker';
-import { MarkdownModule } from 'ngx-markdown';
-import { NgxSkeletonLoaderModule } from 'ngx-skeleton-loader';
+import { provideMarkdown } from 'ngx-markdown';
+import { provideNgxSkeletonLoader } from 'ngx-skeleton-loader';
 import { NgxStripeModule, STRIPE_PUBLISHABLE_KEY } from 'ngx-stripe';
 
 import { environment } from '../environments/environment';
@@ -50,13 +50,11 @@ export function NgxStripeFactory(): string {
     GfLogoComponent,
     GfNotificationModule,
     GfSubscriptionInterstitialDialogModule,
-    MarkdownModule.forRoot(),
     MatAutocompleteModule,
     MatChipsModule,
     MatNativeDateModule,
     MatSnackBarModule,
     MatTooltipModule,
-    NgxSkeletonLoaderModule,
     NgxStripeModule.forRoot(environment.stripePublicKey),
     ServiceWorkerModule.register('ngsw-worker.js', {
       enabled: environment.production,
@@ -68,6 +66,8 @@ export function NgxStripeFactory(): string {
     httpResponseInterceptorProviders,
     LanguageService,
     provideHttpClient(withInterceptorsFromDi()),
+    provideMarkdown(),
+    provideNgxSkeletonLoader(),
     {
       provide: DateAdapter,
       useClass: CustomDateAdapter,