+++ apps/client/src/app/pages/about/changelog/changelog-page.component.ts
@@ -1,11 +1,13 @@
 import { Component, OnDestroy } from '@angular/core';
+import { MarkdownModule } from 'ngx-markdown';
+import { NgxSkeletonLoaderModule } from 'ngx-skeleton-loader';
 import { Subject } from 'rxjs';
 
 @Component({
+  imports: [MarkdownModule, NgxSkeletonLoaderModule],
   selector: 'gf-changelog-page',
   styleUrls: ['./changelog-page.scss'],
-  templateUrl: './changelog-page.html',
-  standalone: false
+  templateUrl: './changelog-page.html'
 })
 export class ChangelogPageComponent implements OnDestroy {
   public isLoading = true;